refactor(harness): extract retained workspace scope owners (SAP-3091)
