+++ packages/harness/src/index.ts
@@ -220,7 +220,7 @@ export {
   PROJECT_AGENT_PROMPT_APPENDIX,
   projectAgentPromptAppendix,
 } from "./profiles/project-agent.js";
-export type { WorkspaceScopeSummary } from "./shared/system-graph.js";
+export type { WorkspaceScopeSummary } from "./shared/workspace-scope.js";
 export { AGENT_STUDIO_PRODUCT_NAME } from "./shared/branding.js";
 export {
   HarnessError,
