+++ packages/harness/src/server/rest.ts
@@ -30,7 +30,7 @@ import type {
   WorkflowInfo,
   SessionInputSubmissionResult,
 } from "../shared/types.js";
-import type { WorkspaceScopeSummary } from "../shared/system-graph.js";
+import type { WorkspaceScopeSummary } from "../shared/workspace-scope.js";
 import type { StudioProjectSummary } from "../shared/agent-map.js";
 import {
   CREATE_SESSION_JSON_LIMIT_BYTES,
