+++ packages/harness/web/src/App.tsx
@@ -59,7 +59,7 @@ import type {
   WorkflowInfo,
   WorkflowInputContractResponse,
 } from "@shared/types";
-import type { WorkspaceKey } from "@shared/system-graph";
+import type { WorkspaceKey } from "@shared/workspace-scope";
 import type {
   StudioProjectId,
   StudioWorkspaceSelection,
