+++ packages/harness/web/src/lib/navigation-history.ts
@@ -1,7 +1,7 @@
 import { useCallback, useRef, useState } from "react";
 
 import type { SessionSummary } from "@shared/types";
-import type { WorkspaceKey } from "@shared/system-graph";
+import type { WorkspaceKey } from "@shared/workspace-scope";
 import type { StudioProjectId } from "@shared/agent-map";
 
 /**
