+++ packages/harness/web/tsconfig.json
@@ -15,7 +15,7 @@
     "paths": {
       "@shared/initial-prompt": ["../src/shared/initial-prompt.ts"],
       "@shared/types": ["../src/shared/types.ts"],
-      "@shared/system-graph": ["../src/shared/system-graph.ts"],
+      "@shared/workspace-scope": ["../src/shared/workspace-scope.ts"],
       "@shared/agent-map-initialization": ["../src/shared/agent-map-initialization.ts"],
       "@shared/agent-map": ["../src/shared/agent-map.ts"],
       "@shared/agent-map-codec": ["../src/shared/agent-map-codec.ts"],
